virtio-mem: MemoryConfig: Add virtiomem_size
Size of virtio-mem is added to MemoryConfig.size as normal memory.
When guest kernel reboot, size of virtio-mem will be setup as normal
memory.
Then the size of virtio-mem cannot decrease after reboot.

This commit add virtiomem_size to MemoryConfig and set it to Mem.new
of virtio-mem to handle the issue.

Fixes: #1593

Signed-off-by: Hui Zhu <teawater@antfin.com>