virtio-mem: Show memory actual size in vm.info
When virtio-mem get resize quest from interface, it will send the
request to the virtio-mem driver in guest kernel.  The workqueue of the
virtio-mem will try to find the suitable memory block and online or
offlint them to resize the memory of virtio-mem and send back
the memory block plug or unplug event to the CLH.
Then the virtio-mem change the memory size is asynchronous.

This commit show memory actual size of virtio-mem in memory_actual_size
of vm.info and add zones_actual_size to vm.info to show memory actual
size of each zone.

Signed-off-by: Hui Zhu <teawater@antfin.com>